+++ HelpmeApp/src/main/java/com/application/helpme/security/services/UserPrinciple.java
@@ -14,7 +14,7 @@ import java.util.stream.Collectors;
 public class UserPrinciple implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private int id;
 
     private String name;
 
@@ -27,7 +27,7 @@ public class UserPrinciple implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserPrinciple(Long id, String name, 
+    public UserPrinciple(int id, String name, 
 			    		String username, String email, String password, 
 			    		Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
@@ -53,7 +53,7 @@ public class UserPrinciple implements UserDetails {
         );
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
